Add files via upload
This is the final code, added LCD stuff, it works on the bench I will test it once I get it all wired up and installed on mu lathe. There is a minor bug in the inches portion of the protocol decoding, I only use mm's so will look at it when I get time.